nasbackend/requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
cffi==1.17.1
cryptography==43.0.3
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
mysqlclient==2.2.6
paramiko==3.5.0
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.2
